[MA-271124] Update AchivementsModel
