+++ Community/Models/Nationalities.swift
@@ -1,0 +1,61 @@
+import Foundation
+
+enum Nationality: String, CaseIterable {
+  case afghan = "Afghan"
+  case albanian = "Albanian"
+  case andorran = "Andorran"
+  case austrian = "Austrian"
+  case bahraini = "Bahraini"
+  case belarusian = "Belarusian"
+  case belgian = "Belgian"
+  case bosnianHerzegovinian = "Bosnian and Herzegovinian"
+  case bulgarian = "Bulgarian"
+  case croatian = "Croatian"
+  case cypriot = "Cypriot"
+  case czech = "Czech"
+  case danish = "Danish"
+  case estonian = "Estonian"
+  case finnish = "Finnish"
+  case french = "French"
+  case german = "German"
+  case greek = "Greek"
+  case hungarian = "Hungarian"
+  case icelandic = "Icelandic"
+  case iranian = "Iranian"
+  case iraqi = "Iraqi"
+  case irish = "Irish"
+  case israeli = "Israeli"
+  case italian = "Italian"
+  case jordanian = "Jordanian"
+  case kazakhstani = "Kazakhstani"
+  case kuwaiti = "Kuwaiti"
+  case kosovar = "Kosovar"
+  case latvian = "Latvian"
+  case lebanese = "Lebanese"
+  case lithuanian = "Lithuanian"
+  case maltese = "Maltese"
+  case moldovan = "Moldovan"
+  case montenegrin = "Montenegrin"
+  case dutch = "Dutch"
+  case northMacedonian = "North Macedonian"
+  case norwegian = "Norwegian"
+  case omani = "Omani"
+  case polish = "Polish"
+  case portuguese = "Portuguese"
+  case qatari = "Qatari"
+  case romanian = "Romanian"
+  case russian = "Russian"
+  case saudiArabian = "Saudi Arabian"
+  case serbian = "Serbian"
+  case slovak = "Slovak"
+  case slovenian = "Slovenian"
+  case spanish = "Spanish"
+  case swedish = "Swedish"
+  case swiss = "Swiss"
+  case syrian = "Syrian"
+  case turkish = "Turkish"
+  case ukrainian = "Ukrainian"
+  case emirati = "Emirati"
+  case british = "British"
+  case yemeni = "Yemeni"
+}
