+++ Community/Models/UserModel.swift
@@ -1,77 +1,6 @@
 import Foundation
 import SwiftUI
 
-struct CountryAchievement: Identifiable, Equatable {
-  let id = UUID()
-  let country: Asset
-  var isUnlocked: Bool
-  var unlockedDate: Date?
-  
-  static func == (lhs: CountryAchievement, rhs: CountryAchievement) -> Bool {
-    lhs.id == rhs.id
-  }
-}
-
-enum Nationality: String, CaseIterable {
-  case afghan = "Afghan"
-  case albanian = "Albanian"
-  case andorran = "Andorran"
-  case austrian = "Austrian"
-  case bahraini = "Bahraini"
-  case belarusian = "Belarusian"
-  case belgian = "Belgian"
-  case bosnianHerzegovinian = "Bosnian and Herzegovinian"
-  case bulgarian = "Bulgarian"
-  case croatian = "Croatian"
-  case cypriot = "Cypriot"
-  case czech = "Czech"
-  case danish = "Danish"
-  case estonian = "Estonian"
-  case finnish = "Finnish"
-  case french = "French"
-  case german = "German"
-  case greek = "Greek"
-  case hungarian = "Hungarian"
-  case icelandic = "Icelandic"
-  case iranian = "Iranian"
-  case iraqi = "Iraqi"
-  case irish = "Irish"
-  case israeli = "Israeli"
-  case italian = "Italian"
-  case jordanian = "Jordanian"
-  case kazakhstani = "Kazakhstani"
-  case kuwaiti = "Kuwaiti"
-  case kosovar = "Kosovar"
-  case latvian = "Latvian"
-  case lebanese = "Lebanese"
-  case lithuanian = "Lithuanian"
-  case maltese = "Maltese"
-  case moldovan = "Moldovan"
-  case montenegrin = "Montenegrin"
-  case dutch = "Dutch"
-  case northMacedonian = "North Macedonian"
-  case norwegian = "Norwegian"
-  case omani = "Omani"
-  case polish = "Polish"
-  case portuguese = "Portuguese"
-  case qatari = "Qatari"
-  case romanian = "Romanian"
-  case russian = "Russian"
-  case saudiArabian = "Saudi Arabian"
-  case serbian = "Serbian"
-  case slovak = "Slovak"
-  case slovenian = "Slovenian"
-  case spanish = "Spanish"
-  case swedish = "Swedish"
-  case swiss = "Swiss"
-  case syrian = "Syrian"
-  case turkish = "Turkish"
-  case ukrainian = "Ukrainian"
-  case emirati = "Emirati"
-  case british = "British"
-  case yemeni = "Yemeni"
-}
-
 struct UserModel {
   var id: Int
   var name: String
@@ -80,19 +9,22 @@ struct UserModel {
   var location: String
   var bio: String
   var profileImage: UIImage?
-  var unlockedCountries: [CountryAchievement] = []
+  var unlockedCountries: [CountryAchievementModel] = []
+}
 
+// MARK: Helpers
+extension UserModel {
   func hasUnlockedCountry(_ country: Asset) -> Bool {
-      unlockedCountries.contains { $0.country == country && $0.isUnlocked }
+    unlockedCountries.contains { $0.country == country && $0.isUnlocked }
   }
 
   func getUnlockDate(for country: Asset) -> Date? {
-      unlockedCountries.first { $0.country == country }?.unlockedDate
+    unlockedCountries.first { $0.country == country }?.unlockedDate
   }
 
   mutating func unlockCountry(_ country: Asset) {
     if !hasUnlockedCountry(country) {
-      let achievement = CountryAchievement(
+      let achievement = CountryAchievementModel(
         country: country,
         isUnlocked: true,
         unlockedDate: Date()
@@ -100,7 +32,7 @@ struct UserModel {
       unlockedCountries.append(achievement)
     }
   }
-  
+
   static func initialUser() -> UserModel {
     UserModel(
       id: 1,
